added documentation refactored code
