+++ src/abstractProof/AbstractPremise.java
@@ -1,16 +1,20 @@
 package abstractProof;
 
-import formulanew.Sentence;
+import formula.Sentence;
 import org.apache.commons.lang3.mutable.MutableInt;
 
 import java.util.ArrayList;
 
+/**
+ * This class implements the abstract representation of a premise. It consists of a single well-formed-formula sentence.
+ * It implements the AbstractStep interface.
+ */
 public final class AbstractPremise implements AbstractStep {
     private final Sentence sentence;
 
 
     public AbstractPremise(Sentence sentence) {
-        this.sentence =sentence;
+        this.sentence = sentence;
     }
 
     @Override
