+++ src/abstractProof/StepRange.java
@@ -2,6 +2,11 @@ package abstractProof;
 
 import org.apache.commons.lang3.Range;
 
+
+/**
+ * This class implements a list of ranges and common operations used for this data structure.
+ * It is used to represent ranges of steps in cited steps of inferences.
+ */
 public final class StepRange {
     private final Range<Integer> range;
