+++ src/formula/AtomicSentence.java
@@ -1,0 +1,7 @@
+package formula;
+
+/**
+ * This interface represents the <atomic formula> non-terminal as specified in the BNF grammar (final report).
+ */
+interface AtomicSentence extends Sentence {
+}
