+++ src/formula/BiImplication.java
@@ -1,18 +1,18 @@
-package formulanew;
+package formula;
 
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 import proof.Operator;
 
-public final class Implication implements Sentence {
+public class BiImplication implements Sentence {
     private final Sentence antecedent;
     private final Sentence consequent;
     private final int hash;
 
-    public Implication(Sentence antecedent, Sentence consequent) {
+    public BiImplication(Sentence antecedent, Sentence consequent) {
         this.antecedent = antecedent;
         this.consequent = consequent;
-        hash = new HashCodeBuilder(9781, 17137).
+        hash = new HashCodeBuilder(2309, 5843).
                 append(this.antecedent).
                 append(this.consequent).
                 toHashCode();
@@ -20,7 +20,7 @@ public final class Implication implements Sentence {
 
     @Override
     public String toString() {
-        return "(" + antecedent + " " + Operator.LIF.getUTFCode() + " " + consequent + ")";
+        return "(" + antecedent + " " + Operator.LIFF.getUTFCode() + " " + consequent + ")";
     }
 
     @Override
@@ -30,15 +30,15 @@ public final class Implication implements Sentence {
 
     @Override
     public boolean equals(Object obj) {
-        if (!(obj instanceof Implication))
+        if (!(obj instanceof BiImplication))
             return false;
         if (obj == this)
             return true;
 
-        Implication toCheck = (Implication) obj;
+        BiImplication other = (BiImplication) obj;
         return new EqualsBuilder().
-                append(this.antecedent, toCheck.antecedent).
-                append(this.consequent, toCheck.consequent).
+                append(this.antecedent, other.antecedent).
+                append(this.consequent, other.consequent).
                 isEquals();
     }
 
@@ -52,10 +52,11 @@ public final class Implication implements Sentence {
 
     @Override
     public boolean isEqualWithReplacement(Sentence other, Argument argument, Argument newArgument) {
-        if (!(other instanceof Implication))
+        if (!(other instanceof BiImplication))
             return false;
         if (other == this)
             return true;
-        return antecedent.isEqualWithReplacement(((Implication) other).antecedent, argument, newArgument) && consequent.isEqualWithReplacement(((Implication) other).consequent, argument, newArgument);
+
+        return antecedent.isEqualWithReplacement(((BiImplication) other).antecedent, argument, newArgument) && consequent.isEqualWithReplacement(((BiImplication) other).consequent, argument, newArgument);
     }
 }
