+++ src/formula/Negation.java
@@ -1,4 +1,4 @@
-package formulanew;
+package formula;
 
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
