--- src/formula/Predicate.java
@@ -1,7 +1,0 @@
-package formula;
-
-import proof.Operator;
-
-public interface Predicate {
-    Operator getMainOperator();
-}
