+++ src/formula/PredicateSymbol.java
@@ -2,13 +2,13 @@ package formula;
 
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
-import proof.Operator;
+
 
 import java.util.ArrayList;
 import java.util.stream.Collectors;
 
 
-public final class PredicateSymbol implements Predicate {
+public final class PredicateSymbol implements AtomicSentence {
     private final String name;
     private final ArrayList<Argument> arguments;
     private final int hash;
@@ -24,8 +24,7 @@ public final class PredicateSymbol implements Predicate {
 
     @Override
     public String toString() {
-        return name + ((arguments != null) ? ("(" + arguments.stream().limit(arguments.size() - 1).map(s -> s + ", ").collect(Collectors.joining())
-                + arguments.get(arguments.size() - 1) + ")") : "");
+        return name + ((arguments != null) ? ("(" + arguments.stream().limit(arguments.size() - 1).map(s -> s + ", ").collect(Collectors.joining()) + arguments.get(arguments.size() - 1) + ")") : "");
     }
 
     @Override
@@ -35,10 +34,8 @@ public final class PredicateSymbol implements Predicate {
 
     @Override
     public boolean equals(Object obj) {
-        if (!(obj instanceof PredicateSymbol))
-            return false;
-        if (obj == this)
-            return true;
+        if (!(obj instanceof PredicateSymbol)) return false;
+        if (obj == this) return true;
 
         PredicateSymbol toCheck = (PredicateSymbol) obj;
         return new EqualsBuilder().
@@ -48,7 +45,20 @@ public final class PredicateSymbol implements Predicate {
     }
 
     @Override
-    public Operator getMainOperator() {
-        return null;
+    public boolean isEqualWithReplacement(Sentence other, Argument argument, Argument newArgument) {
+        if (!(other instanceof PredicateSymbol)) return false;
+        if (other == this) return true;
+        PredicateSymbol otherPredicateSymbol = (PredicateSymbol) other;
+
+        if (!this.name.equals(otherPredicateSymbol.name)) {
+            return false;
+        }
+
+        for (int idx = 0; idx < arguments.size(); ++idx) {
+            if (!(arguments.get(idx).isEqualWithReplacement(otherPredicateSymbol.arguments.get(idx), argument, newArgument))) {
+                return false;
+            }
+        }
+        return true;
     }
 }
