+++ src/formula/Quantifier.java
@@ -1,7 +1,4 @@
 package formula;
 
-import proof.Operator;
-
 public interface Quantifier {
-    Operator getMainOperator();
 }
