--- src/formulanew/Argument.java
@@ -1,5 +1,0 @@
-package formulanew;
-
-public interface Argument {
-    public boolean isEqualWithReplacement(Argument other, Argument argument, Argument newArgument);
-}
