--- src/formulanew/AtomicSentence.java
@@ -1,4 +1,0 @@
-package formulanew;
-
-interface AtomicSentence extends Sentence {
-}
