--- src/formulanew/Quantifier.java
@@ -1,4 +1,0 @@
-package formulanew;
-
-public interface Quantifier {
-}
