--- src/formulanew/Sentence.java
@@ -1,6 +1,0 @@
-package formulanew;
-
-
-public interface Sentence {
-    public boolean isEqualWithReplacement(Sentence other, Argument argument, Argument newArgument);
-}
