+++ src/tokenizer/FormulaTokenizer.java
@@ -6,6 +6,11 @@ import java.util.Arrays;
 import java.util.StringTokenizer;
 import java.util.stream.Collectors;
 
+
+/**
+ * This class extends the functionality of StringTokenizer.
+ * Particularly, a token can be inspected without advancing the pointer.
+ */
 public class FormulaTokenizer extends StringTokenizer {
     private static final String skipdelims = " ";
     private static final String delims = "(),"+ Arrays.stream(Operator.values()).map(Operator::getUTFCode).collect(Collectors.joining());
@@ -33,6 +38,12 @@ public class FormulaTokenizer extends StringTokenizer {
         return tkn != null;
     }
 
+    /**
+     * Verifies if the next token is equal to the specified string.
+     * If it is the case the token pointer is advanced.
+     * @param string the next token
+     * @return true if token is equal, otherwise false.
+     */
     public boolean isNextEqualTo(String string) {
         if (tkn == null || !tkn.equals(string)) {
             return false;
